fixes path setup (for windows binary)
path setup has to happen before _any_ import from ilastik.

(ilastik/__init__.py already imports external packages).
